rename to receive
